requirements.txt
# =============================================================================
# RIADD Modern - Requirements
# =============================================================================
# Tested and pinned versions for stability
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Deep Learning
# -----------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# -----------------------------------------------------------------------------
# Image Augmentation
# -----------------------------------------------------------------------------
albumentations>=1.3.0

# -----------------------------------------------------------------------------
# Machine Learning (for ensemble)
# -----------------------------------------------------------------------------
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
PyYAML>=6.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0

# -----------------------------------------------------------------------------
# Progress & Logging
# -----------------------------------------------------------------------------
tqdm>=4.65.0
rich>=13.0.0

# -----------------------------------------------------------------------------
# Experiment Tracking (for monitoring overfitting)
# -----------------------------------------------------------------------------
wandb>=0.15.0

# -----------------------------------------------------------------------------
# Scipy (for entropy calculations in ChaosFEX)
# -----------------------------------------------------------------------------
scipy>=1.10.0

# -----------------------------------------------------------------------------
# Jupyter Support (optional, for notebooks)
# -----------------------------------------------------------------------------
ipykernel>=6.0.0
ipywidgets>=8.0.0
